+++ contracts/service-provider-directory/src/types.rs
@@ -1,3 +1,8 @@
+use cosmwasm_std::{Addr, Coin, Storage};
+use cw_storage_plus::{Index, IndexList, IndexedMap, Item, MultiIndex};
+use serde::{Deserialize, Serialize};
+
+use crate::error::Result;
 
 /// The directory of services are indexed by [`ServiceId`].
 pub type ServiceId = u32;
